// rafts/frontend/src/services/types.ts
export interface StorageError {
  status: number
  message: string
}

export interface StorageResponse<T> {
  data?: T
  error?: StorageError
}